extract swtich
