+++ server/src/domain/checks/check.service.ts
@@ -6,7 +6,7 @@ import type { MonitorPayloadMap, MonitorStatusResponse } from "@/types/network.j
 import type { HardwareStatusPayload, PageSpeedStatusPayload } from "@/types/network.js";
 import { AppError } from "@/utils/AppError.js";
 import { ILogger } from "@/utils/logger.js";
-import { DateRange } from "@/types/query.js";
+import { CheckFilter, DateRange } from "@/types/query.js";
 
 const SERVICE_NAME = "checkService";
 
@@ -22,7 +22,7 @@ export interface ICheckService {
 		dateRange: DateRange;
 		page: number;
 		rowsPerPage: number;
-		filter?: string;
+		filter?: CheckFilter;
 		status?: boolean;
 	}): Promise<ChecksQueryResult>;
 	getChecksByTeam(params: {
@@ -31,7 +31,7 @@ export interface ICheckService {
 		dateRange: DateRange;
 		page: number;
 		rowsPerPage: number;
-		filter?: string;
+		filter?: CheckFilter;
 	}): Promise<ChecksQueryResult>;
 	getChecksSummaryByTeamId(params: { teamId: string; dateRange: DateRange }): Promise<ChecksSummary>;
 	deleteChecks(params: { monitorId: string; teamId: string }): Promise<number>;
@@ -169,7 +169,7 @@ export class CheckService implements ICheckService {
 		page: number;
 		rowsPerPage: number;
 		status?: boolean;
-		filter?: string;
+		filter: CheckFilter;
 	}) => {
 		if (!monitorId) {
 			throw new AppError({ message: "No monitor ID in request", service: SERVICE_NAME, method: "getChecksByMonitor", status: 400 });
@@ -184,7 +184,7 @@ export class CheckService implements ICheckService {
 		const parsedPage = page ?? 0;
 		const parsedRowsPerPage = rowsPerPage ?? 5;
 
-		const result = await this.checksRepository.findByMonitorId(monitorId, sortOrder, dateRange, filter, parsedPage, parsedRowsPerPage, status);
+		const result = await this.checksRepository.findByMonitorId(monitorId, sortOrder, dateRange, parsedPage, parsedRowsPerPage, status, filter);
 
 		return result;
 	};
@@ -193,21 +193,21 @@ export class CheckService implements ICheckService {
 		teamId,
 		sortOrder,
 		dateRange,
-		filter,
 		page,
 		rowsPerPage,
+		filter,
 	}: {
 		teamId: string;
 		sortOrder: string;
 		dateRange: DateRange;
 		page: number;
 		rowsPerPage: number;
-		filter?: string;
+		filter?: CheckFilter;
 	}) => {
 		const parsedPage = page ?? 0;
 		const parsedRowsPerPage = rowsPerPage ?? 5;
 
-		const checkData = await this.checksRepository.findByTeamId(sortOrder, dateRange, filter, parsedPage, parsedRowsPerPage, teamId);
+		const checkData = await this.checksRepository.findByTeamId(sortOrder, dateRange, parsedPage, parsedRowsPerPage, teamId, filter);
 		return checkData;
 	};
