+++ server/src/service/network/GameProvider.ts
@@ -3,7 +3,7 @@ import { GameStatusPayload, MonitorStatusResponse } from "@/types/network.js";
 import { Monitor, MonitorType } from "@/domain/monitors/monitor.types.js";
 import { GameDig } from "gamedig";
 import { AppError } from "@/utils/AppError.js";
-import { NETWORK_ERROR } from "@/service/network/utils.js";
+import { NETWORK_ERROR } from "@/types/network.js";
 import { ILogger } from "@/utils/logger.js";
 
 type GameDigType = typeof GameDig;
