+++ server/src/service/network/HttpProvider.ts
@@ -6,7 +6,8 @@ import { MonitorStatusResponse } from "@/types/network.js";
 import { Agent as HttpsAgent } from "https";
 import { Agent as HttpAgent } from "http";
 import { Monitor, MonitorType } from "@/domain/monitors/monitor.types.js";
-import { NETWORK_ERROR, isStatusUp } from "@/service/network/utils.js";
+import { isStatusUp } from "@/service/network/utils.js";
+import { NETWORK_ERROR } from "@/types/network.js";
 import CacheableLookup from "cacheable-lookup";
 
 export class HttpProvider implements IStatusProvider<HttpStatusPayload> {
