+++ server/test/unit/services/checkService.test.ts
@@ -375,7 +375,7 @@ describe("CheckService", () => {
 				rowsPerPage: null as any,
 			});
 
-			expect(checksRepository.findByMonitorId).toHaveBeenCalledWith("mon-1", "desc", "day", undefined, 0, 5, undefined);
+			expect(checksRepository.findByMonitorId).toHaveBeenCalledWith("mon-1", "desc", "day", 0, 5, undefined, undefined);
 		});
 
 		it("passes filter and status parameters", async () => {
@@ -388,11 +388,11 @@ describe("CheckService", () => {
 				dateRange: "week",
 				page: 2,
 				rowsPerPage: 20,
-				filter: "error",
+				filter: "resolve",
 				status: false,
 			});
 
-			expect(checksRepository.findByMonitorId).toHaveBeenCalledWith("mon-1", "asc", "week", "error", 2, 20, false);
+			expect(checksRepository.findByMonitorId).toHaveBeenCalledWith("mon-1", "asc", "week", 2, 20, false, "resolve");
 		});
 
 		it("throws when monitorId is missing", async () => {
@@ -443,7 +443,7 @@ describe("CheckService", () => {
 
 			await service.getChecksByTeam({ teamId: "team-1", sortOrder: "desc", dateRange: "day", page: null as any, rowsPerPage: null as any });
 
-			expect(checksRepository.findByTeamId).toHaveBeenCalledWith("desc", "day", undefined, 0, 5, "team-1");
+			expect(checksRepository.findByTeamId).toHaveBeenCalledWith("desc", "day", 0, 5, "team-1", undefined);
 		});
 	});
